refactor: move builtin player related models to a separate module
